pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tencent-valuation-v4"
version = "0.4.1"
description = "Tencent (0700.HK) valuation pipeline V4 with multi-method stack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "PyYAML>=6.0",
  "statsmodels>=0.14",
  "beautifulsoup4>=4.12",
  "pypdf>=4.2"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
tencent-valuation-v4 = "tencent_valuation_v4.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
